backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.31
aiosqlite==0.20.0
pydantic==2.7.1
pydantic-settings==2.3.2
openpyxl==3.1.2
python-docx==1.1.0
openai==1.30.5
aiohttp==3.9.5
pytest==8.2.1
pytest-asyncio==0.23.7